refine names
